apps/api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Authentication
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Email (SMTP)
aiosmtplib>=3.0.1

# Caching
fastapi-cache2==0.2.2
redis>=5.0.0

# Encryption
cryptography>=41.0.0

# Telegram integration
telethon>=1.33.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
faker>=40.0.0

# Templates and static files
jinja2>=3.1.2
python-multipart>=0.0.9

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
locust>=2.43.0

# WebSockets
websockets>=12.0

# Utilities
python-dateutil>=2.8.2

# HTTP client (async)
httpx>=0.27.0
